pyproject.toml
[tool.poetry]
name = "atomicprogress-by-evoq"
version = "0.1.0"
description = "Atomic Progress Tracker is your ultimate companion on the journey to building life-changing habits inspired by James Clear's \"Atomic Habits.\" It empowers you to break down your goals into small, manageable atomic habits. With our intuitive tracking system, you can monitor your daily progress, gain insights into your habits, and make consistent improvements over time."
authors = ["Tigran Saatchyan <mr.saatchyan@gmail.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.6"
psycopg = "^3.1.12"
djangorestframework = "^3.14.0"
Pillow = "^10.0.1"
djangorestframework-simplejwt = "^5.3.0"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.2.0"
celery = "^5.3.4"
django-celery-beat = "^2.5.0"
redis = "^5.0.1"
django-filter = "^23.3"
maskpass = "^0.3.7"
requests = "^2.31.0"
pytelegrambotapi = "^4.14.0"

[tool.poetry.group.develop.dependencies]
python-dotenv = "^1.0.0"
flake8 = "^6.1.0"
flake8-django = "^1.4"
pre-commit = "^3.4.0"
coverage = "^7.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
